Change Architecture / save-load to file / list room-guest / update TextUI

+++ src/pl/chylu/App.java
@@ -7,7 +7,7 @@ public class App {
 
     public static void main(String[] args) {
         String hotelName = "TimeToSleep";
-        double systemVersion = 1.0;
+        int systemVersion = 1;
         boolean isDeveloperVersion = true;
 
         textUI.showSystemInfo(hotelName, systemVersion, isDeveloperVersion);
